requirements.txt
streamlit
pandas
plotly
openpyxl
plotly_express
streamlit-authenticator
pyyaml
